pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=63.0.0"]

[project]
name = "apify-aal-utils"
version = "0.1.0"
dependencies = [
  "google-cloud-bigquery >=3.16.0, <4.0.0",
]
requires-python = ">=3.8"
readme = "README.rst"
description = "Reusable utilities for Apify AAL Actors"

[tool.isort]
profile = "black"
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
